requirements.txt
pdfplumber
python-docx
requests
beautifulsoup4
spacy
sentence-transformers
